Fix areas by product
Previously the area field on Radar was only used for iOS. Now it's used
for multiple products, yet the items with the same names across products
don't particularly have the same IDs. This updates the area picker to be
populated by the areas per product, and cleared when switching to a
product with no areas.