Fix non-responsive menu bar on first launch
After launching brisk with the dock icon active, if you attempted to
click the menu bar immediately after it was launched, it would not be
interactive. To fix this you had to focus a different app, and then
refocus brisk.

This seemed to be caused by the setActivationPolicy call being in
applicationDidFinishLaunching rather than
applicationWillFinishLaunching. Moving this initialization there seems to
fix this issue.

https://stackoverflow.com/questions/33345686/cocoa-application-menu-bar-not-clickable