Create custom TextView class to set font
Previously when you typed an emoji into one of the NSTextViews it would
cause the font to change. This seems to be because there was no selected
font, so something breaks internally, causing it to change to a
monospace font. To fix this we just need to explicitly set a font on the
TextView. We would do this from IB except that only seems to stick if
you also provide default text for the text view.