Use single user icon for open radar preferences
For now this can be used as the last icon for preferences. In the future
I think we should change to a more Mail.app like approach where we have
a single accounts tab that has Apple Radar, Open Radar, and maybe things
like the Swift JIRA and Twitter. In that case we'll only have a single
icon.