Only update title with valid title text
Right now whenever you type anything in any field we wipe out the title
field. We should only update the title when the title text is non-empty.
I would set the document's title to nil when you delete it, but if you
do that you end up with a document named "Untitled -1" which then syncs
with the window's title bar and seems crazy.