Reset change count when loading radar
Previously if you started to dup a radar, we filled in all the fields in
the window, which sets the document change count to edited. Then if you
closed the window, you would be prompted to save the document. In this
case where you haven't edited anything, the document will be immediately
closable and you will only be prompted to save the document if you make
changes to it.