Add checkbox for crossposting to open radar
This adds a checkbox in each window for whether or not the current radar
should be cross posted to open radar. This is useful for when you want
to keep a radar private, or you just don't want to cross post a
duplicate to open radar.

This won't scale if we want to allow:

- only posting to openradar (perhaps if you just want to post something
you've already posted to apple radar)
- support for more services

But in the meantime I think having a way to customize this is important
so it's worth adding this way for now.